package duplicate name fix nameCreate duplicate logic fix to directly name, index null check readme fix

+++ src/main/kotlin/com/example/berry_server/controller/NameController.kt
@@ -1,10 +1,11 @@
 package com.example.berry_server.controller
 
 import com.example.berry_server.util.Constants
-import com.example.berry_server.dto.model.NameItem
 import com.example.berry_server.dto.request.name.CreateNameRequest
 import com.example.berry_server.dto.request.name.DeleteNameRequest
+import com.example.berry_server.dto.response.ApiErrorResponse
 import com.example.berry_server.dto.response.ApiResponse
+import com.example.berry_server.dto.response.BaseApiResponse
 import com.example.berry_server.repository.NameRepository
 import com.example.berry_server.util.Messages
 import com.example.berry_server.util.Validation
@@ -23,10 +24,18 @@ class NameController(
 
     // 이름 생성
     @PostMapping("/createName")
-    fun createName(@RequestBody request: CreateNameRequest): ApiResponse<List<NameItem>> {
+    fun createName(@RequestBody request: CreateNameRequest): BaseApiResponse {
+        // null 체크
+        if (request.name == null) {
+            return ApiErrorResponse(
+                message = Constants.ERROR,
+                error = Messages.ERROR_CREATE_NAME_INPUT_NAME
+            )
+        }
 
+        // 이름 validation
         Validation.validateName(request.name)?.let { errorMessage ->
-            return ApiResponse(
+            return ApiErrorResponse(
                 message = Constants.ERROR,
                 error = errorMessage
             )
@@ -34,9 +43,9 @@ class NameController(
 
         // 중복 검사
         if (nameRepository.existName(request.name)) {
-            return ApiResponse(
-                error = Constants.ERROR,
-                message = Messages.ERROR_CREATE_NAME_DUPLICATED
+            return ApiErrorResponse(
+                message = Constants.ERROR,
+                error = Messages.ERROR_CREATE_NAME_DUPLICATED
             )
         }
 
@@ -50,7 +59,7 @@ class NameController(
 
     // 이름 전체 조회
     @GetMapping("/getName")
-    fun getName(): ApiResponse<List<NameItem>> {
+    fun getName(): BaseApiResponse {
         return ApiResponse(
             message = Constants.SUCCESS,
             data = nameRepository.getNameList()
@@ -59,12 +68,25 @@ class NameController(
 
     // 인덱스 이름 삭제
     @DeleteMapping("/deleteName")
-    fun deleteName(@RequestBody request: DeleteNameRequest): ApiResponse<List<NameItem>> {
+    fun deleteName(@RequestBody request: DeleteNameRequest): BaseApiResponse {
+        if (request.deleteIndex == null) {
+            return ApiErrorResponse(
+                message = Constants.ERROR,
+                error = Messages.ERROR_DELETE_NAME_INDEX_NULL,
+            )
+        }
+
         val isDeleted = nameRepository.deleteName(request.deleteIndex)
 
+        if (!isDeleted) {
+            return ApiErrorResponse(
+                message = Constants.ERROR,
+                error = "${Messages.ERROR_DELETE_NAME_INVALID_INDEX}: ${request.deleteIndex}",
+            )
+        }
+
         return ApiResponse(
-            message = if (isDeleted) Constants.SUCCESS else Constants.ERROR,
-            error = if (isDeleted) null else "${Messages.ERROR_DELETE_NAME_INVALID_INDEX}: ${request.deleteIndex}",
+            message = Constants.SUCCESS,
             data = nameRepository.getNameList()
         )
     }
